+++ CHANGES.rst
@@ -12,4 +12,4 @@ Change log
 1.0 (2017-01-06)
 ================
 
-* Initial release.
+- Initial release.