+++ CHANGES.rst
@@ -2,7 +2,7 @@
 Change log
 ==========
 
-1.1 (unreleased)
+1.1 (2017-01-07)
 ================
 
 - Add ``make-current-addressbook`` script to ease usage of the `current`